Adding backend for showing record

+++ src/main/java/com/example/AttendanceSystem/Repositry/StudentRepositry.java
@@ -3,6 +3,6 @@ package com.example.AttendanceSystem.Repositry;
 import com.example.AttendanceSystem.Entity.Students;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepositry  extends JpaRepository<Students , Integer> {
+public interface StudentRepositry  extends JpaRepository<Students , Long> {
 
 }
